moving into unit tests
