+++ torch_spyre/_inductor/layout_backend.py
@@ -3,8 +3,6 @@ import copy
 from itertools import combinations
 from typing import List, Tuple, Literal
 import numpy as np
-import matplotlib.pyplot as plt
-from matplotlib import patches
 
 
 @dataclass
@@ -42,90 +40,6 @@ class Operation:
         reads = [self._buffer_registry[buffer_name] for buffer_name in self.inputs]
         writes = [self._buffer_registry[buffer_name] for buffer_name in self.outputs]
         return ReadWrites(reads=reads, writes=writes)
-
-
-def plot_layout(capacity: int, buffers: list[Buffer]):
-    assert np.all([b.start_time is not None for b in buffers]), (
-        "Start time must be defined"
-    )
-    assert np.all([b.end_time is not None for b in buffers]), "End time must be defined"
-
-    fig, ax = plt.subplots(figsize=(10, 6))
-    ax.axhline(
-        y=capacity,
-        color="red",
-        linestyle="--",
-        linewidth=2,
-        label=f"SPM Ceiling ({capacity})",
-    )
-    ax.axhline(y=0, color="red", linestyle="--", linewidth=2, label="SPM Floor")
-
-    max_time = max([b.end_time for b in buffers if b.end_time is not None])
-    ax.set_xlim(0, max_time + 1)
-    ax.set_ylim(-capacity, capacity * 2)
-    ax.set_xlabel("Time (Logical Steps)")
-    ax.set_ylabel("Memory Address (Bytes)")
-    ax.set_title("Calculated Memory Layout")
-
-    rects = {}
-    texts = {}
-
-    # Initialize the visual blocks
-    for b in buffers:
-        rect = patches.Rectangle(
-            (b.start_time, 0),
-            b.end_time - b.start_time,
-            b.size,
-            linewidth=1.5,
-            edgecolor="black",
-            facecolor="skyblue",
-            alpha=0.8,
-        )
-        ax.add_patch(rect)
-        rects[b.name] = rect
-
-        txt = ax.text(
-            0,
-            0,
-            f"Buffer {b.name}",
-            ha="center",
-            va="center",
-            fontsize=10,
-            weight="bold",
-        )
-        texts[b.name] = txt
-
-    step_text = ax.text(
-        0.02, 0.90, "", transform=ax.transAxes, fontsize=12, weight="bold"
-    )
-    fig.canvas.draw()
-    fig.canvas.flush_events()
-
-    msg = "FINISHED! Showing Final Configuration"
-    for b in buffers:
-        if not b.spilled:
-            rects[b.name].set_visible(True)
-            texts[b.name].set_visible(True)
-            rects[b.name].set_y(b.address)
-            texts[b.name].set_position(
-                (
-                    b.start_time + (b.end_time - b.start_time) / 2,
-                    b.address + b.size / 2,
-                )
-            )
-        else:
-            rects[b.name].set_visible(False)
-            texts[b.name].set_visible(False)
-
-    # Add a warning box for spilled buffers
-    spilled_buffers = [b.name + f" : {b.density:.2f}" for b in buffers if b.spilled]
-    if spilled_buffers:
-        msg += f"\nSpilled to DRAM: {', '.join(spilled_buffers)}"
-    else:
-        msg += "\n No evictions required"
-
-    step_text.set_text(msg)
-    plt.show()
 
 
 # def calculate_logical_lifetimes(ops: list[Operation]) -> list[Buffer]:
@@ -226,7 +140,7 @@ def allocate_sa(
     overlapping_time: dict[str, list[tuple[str, int]]] = {}
     for buffer in buffers:
         overlapping_time[buffer.name] = []
-        buffer.address = rng.integers(0, capacity - buffer.size)
+        buffer.address = rng.integers(0, capacity - buffer.size) if capacity > buffer.size else 0
 
     normalize_buffer_sizes(buffers)
 
@@ -367,7 +281,7 @@ def allocate_buffers(
         assert b.start_time < b.end_time, "Buffer must start before end"
 
     # Don't try to assign buffers which are bigger than capacity
-    _buffers = [b for b in buffers if b.size < capacity]
+    _buffers = [b for b in buffers if b.size <= capacity]
 
     result: None | List[Buffer] = None
     match method:
@@ -385,19 +299,3 @@ def allocate_buffers(
         # TODO: compare results for optimality if no obvious improvement...
         return fallback_result
     return result
-
-
-# ==========================================
-# Example Usage
-# ==========================================
-if __name__ == "__main__":
-    SPM_CAPACITY = 1000
-
-    buffer_list = [
-        Buffer(name="A", size=750, start_time=0, end_time=3),
-        Buffer(name="B", size=50, start_time=1, end_time=4),
-        Buffer(name="C", size=800, start_time=3, end_time=5),
-    ]
-
-    best_allocation = allocate_buffers(SPM_CAPACITY, buffer_list, "ordered-global")
-    plot_layout(SPM_CAPACITY, best_allocation)
